requirements.txt
anthropic>=0.40.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn>=0.30.0
click>=8.1.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pyyaml>=6.0.0
rich>=13.0.0
httpx>=0.27.0
pytest>=8.0.0
